# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "윤정 법률사무소 API",
    "version": "1.0.0",
    "description": "윤정 법률사무소 홈페이지 백엔드 REST API. 모든 응답은 { data, error, meta } 형식입니다."
  },
  "servers": [
    { "url": "/api", "description": "기본 API 서버" }
  ],
...
  "paths": {
    "/documents": {
      "get": {
        "tags": ["Documents"],
        "summary": "문서 목록 조회",
        "description": "페이지네이션, 문서 유형/상태/중요도 필터, 키워드 검색을 지원합니다.",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 }, "description": "페이지 번호" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 }, "description": "페이지당 항목 수" },
          { "name": "document_type", "in": "query", "schema": { "type": "string" }, "description": "문서 유형 필터" },
          { "name": "status", "in": "query", "schema": { "type": "string" }, "description": "상태 필터 (unread, read, archived 등)" },
          { "name": "importance", "in": "query", "schema": { "type": "string" }, "description": "중요도 필터" },
          { "name": "q", "in": "query", "schema": { "type": "string" }, "description": "키워드 검색" }
        ],
        "responses": {
          "200": {
            "description": "문서 목록",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Document" }
                        },
                        "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Documents"],
        "summary": "문서 생성",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DocumentCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "생성된 문서",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Document" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/documents/search": {
      "get": {
        "tags": ["Documents"],
        "summary": "FTS5 전문 검색",
        "description": "SQLite FTS5를 사용한 한국어 전문 검색입니다.",
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string" }, "description": "검색 키워드" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 }, "description": "최대 결과 수" }
        ],
        "responses": {
          "200": {
            "description": "검색 결과",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Document" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/documents/upload": {
      "post": {
        "tags": ["Documents"],
        "summary": "파일 업로드",
        "description": "PDF, MD, TXT, HTML 파일을 업로드하고 텍스트를 추출합니다. 최대 50MB.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "업로드할 파일 (pdf, md, txt, html)"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "업로드 결과",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "filePath": { "type": "string", "description": "저장된 파일 경로" },
                            "contentPlain": { "type": "string", "nullable": true, "description": "추출된 텍스트" },
                            "fileType": { "type": "string", "description": "파일 유형 (pdf, markdown, text, html)" },
                            "fileSize": { "type": "integer", "description": "파일 크기 (바이트)" }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/documents/upload-markdown": {
      "post": {
        "tags": ["Documents"],
        "summary": "마크다운 업로드 + 자동 분석/문서 생성",
        "description": "마크다운 파일을 업로드하면 자동으로 분석하여 유형, 키워드, 태그를 추출하고 문서를 생성합니다.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "마크다운 또는 텍스트 파일 (md, markdown, txt)"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "생성된 문서 및 분석 결과",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "document": { "$ref": "#/components/schemas/Document" },
                            "analysis": {
                              "type": "object",
                              "properties": {
                                "documentType": { "type": "string" },
                                "keywords": { "type": "array", "items": { "type": "string" } },
                                "suggestedCategories": { "type": "array", "items": { "type": "string" } },
                                "structure": { "type": "object" },
                                "frontmatter": { "type": "object", "nullable": true }
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/documents/analyze-markdown": {
      "post": {
        "tags": ["Documents"],
        "summary": "마크다운 분석 미리보기 (문서 생성 없음)",
        "description": "마크다운 파일 또는 텍스트를 분석하여 결과만 반환합니다. 문서를 생성하지 않습니다.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": { "type": "string", "description": "마크다운 텍스트" },
                  "filename": { "type": "string", "description": "파일명 (분석 힌트용)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "분석 결과",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "type": "object", "description": "분석 결과 객체" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/documents/{id}": {
      "get": {
        "tags": ["Documents"],
        "summary": "문서 상세 조회",
        "description": "태그, 카테고리, 컬렉션, 하이라이트, 관계 정보를 포함한 문서 상세 데이터를 반환합니다.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "description": "문서 ID" }
        ],
        "responses": {
          "200": {
            "description": "문서 상세",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Document" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Documents"],
        "summary": "문서 수정",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "description": "문서 ID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DocumentCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정된 문서",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Document" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Documents"],
        "summary": "문서 삭제",
        "description": "첫 번째 삭제 시 archived 상태로 변경, 두 번째 삭제 시 영구 삭제됩니다.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "description": "문서 ID" }
        ],
        "responses": {
          "200": {
            "description": "삭제 결과",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "관리자 인증 토큰 (POST /admin-users/login에서 발급)"
      },
      "portalAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-portal-token",
        "description": "포털 세션 토큰 (POST /portal/login에서 발급, TTL 24시간)"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "description": "모든 API의 공통 응답 형식",
        "properties": {
          "data": { "description": "응답 데이터 (성공 시)", "nullable": true },
          "error": { "type": "string", "description": "에러 메시지 (실패 시)", "nullable": true },
          "meta": { "type": "object", "description": "부가 정보 (페이지네이션 등)", "nullable": true }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "description": "현재 페이지" },
          "limit": { "type": "integer", "description": "페이지당 항목 수" },
          "total": { "type": "integer", "description": "전체 항목 수" },
          "totalPages": { "type": "integer", "description": "전체 페이지 수" }
        }
      },
      "Document": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "description": "문서 ID" },
          "title": { "type": "string", "description": "제목" },
          "documentType": { "type": "string", "description": "문서 유형" },
          "subtitle": { "type": "string", "nullable": true },
          "author": { "type": "string", "nullable": true },
          "source": { "type": "string", "nullable": true },
          "publishedDate": { "type": "string", "nullable": true },
          "contentMarkdown": { "type": "string", "nullable": true },
          "contentPlain": { "type": "string", "nullable": true },
          "summary": { "type": "string", "nullable": true },
          "status": { "type": "string", "description": "상태 (unread, read, archived)" },
          "importance": { "type": "string", "nullable": true },
          "filePath": { "type": "string", "nullable": true },
          "fileType": { "type": "string", "nullable": true },
          "metadata": { "type": "string", "nullable": true, "description": "JSON 문자열" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "DocumentCreate": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "description": "제목" },
          "documentType": { "type": "string" },
          "subtitle": { "type": "string" },
          "author": { "type": "string" },
          "source": { "type": "string" },
          "publishedDate": { "type": "string" },
          "contentMarkdown": { "type": "string" },
          "contentPlain": { "type": "string" },
          "summary": { "type": "string" },
          "status": { "type": "string" },
          "importance": { "type": "string" },
          "metadata": { "type": "string" }
        },
        "required": ["title"]
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "잘못된 요청",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                { "$ref": "#/components/schemas/ApiResponse" },
                {
                  "type": "object",
                  "properties": {
                    "data": { "type": "null" },
                    "error": { "type": "string", "example": "필수 항목이 누락되었습니다" }
                  }
                }
              ]
            }
          }
        }
      },
...
      "NotFound": {
        "description": "리소스를 찾을 수 없음",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                { "$ref": "#/components/schemas/ApiResponse" },
                {
                  "type": "object",
                  "properties": {
                    "data": { "type": "null" },
                    "error": { "type": "string", "example": "리소스를 찾을 수 없습니다" }
                  }
                }
              ]
            }
          }
        }
      }
    }
  }
}